Move existing golinks and plugin activation tests to the "core" folder.
